+++ sonar-plugin/sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/analysis/TsConfigProvider.java
@@ -19,10 +19,8 @@
  */
 package org.sonar.plugins.javascript.analysis;
 
-import static java.util.Arrays.stream;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
-import static java.util.stream.Stream.concat;
 
 import com.google.gson.Gson;
 import java.io.File;
@@ -38,7 +36,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
-import java.util.function.Predicate;
 import javax.annotation.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -47,12 +44,9 @@ import org.sonar.api.batch.fs.FilePredicate;
 import org.sonar.api.batch.fs.FileSystem;
 import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.batch.sensor.SensorContext;
-import org.sonar.api.config.Configuration;
-import org.sonar.api.utils.WildcardPattern;
+import static org.sonar.plugins.javascript.analysis.LookupConfigProviderFilter.FileFilter;
+import static org.sonar.plugins.javascript.analysis.LookupConfigProviderFilter.PathFilter;
 import org.sonar.plugins.javascript.JavaScriptFilePredicate;
-import org.sonar.plugins.javascript.JavaScriptLanguage;
-import org.sonar.plugins.javascript.JavaScriptPlugin;
-import org.sonar.plugins.javascript.TypeScriptLanguage;
 import org.sonar.plugins.javascript.sonarlint.TsConfigCache;
 import org.sonarsource.analyzer.commons.FileProvider;
 
@@ -234,68 +228,6 @@ public class TsConfigProvider {
 
     public TsConfigOrigin type() {
       return TsConfigOrigin.LOOKUP;
-    }
-
-    static class FileFilter implements Predicate<Path> {
-
-      private final Set<String> extensions = new HashSet<>();
-
-      public FileFilter(Configuration config) {
-        extensions.addAll(
-          Arrays.asList(
-            config
-              .get(JavaScriptLanguage.FILE_SUFFIXES_KEY)
-              .orElse(JavaScriptLanguage.FILE_SUFFIXES_DEFVALUE)
-              .split(",")
-          )
-        );
-        extensions.addAll(
-          Arrays.asList(
-            config
-              .get(TypeScriptLanguage.FILE_SUFFIXES_KEY)
-              .orElse(TypeScriptLanguage.FILE_SUFFIXES_DEFVALUE)
-              .split(",")
-          )
-        );
-      }
-
-      @Override
-      public boolean test(Path path) {
-        return extensions.stream().anyMatch(ext -> path.toString().endsWith(ext));
-      }
-    }
-
-    static class PathFilter implements Predicate<Path> {
-
-      private final WildcardPattern[] exclusions;
-
-      public PathFilter(Configuration config) {
-        if (!isExclusionOverridden(config)) {
-          exclusions = WildcardPattern.create(JavaScriptPlugin.EXCLUSIONS_DEFAULT_VALUE);
-        } else {
-          WildcardPattern[] jsExcludedPatterns = WildcardPattern.create(
-            config.getStringArray(JavaScriptPlugin.JS_EXCLUSIONS_KEY)
-          );
-          WildcardPattern[] tsExcludedPatterns = WildcardPattern.create(
-            config.getStringArray(JavaScriptPlugin.TS_EXCLUSIONS_KEY)
-          );
-          exclusions =
-            concat(stream(jsExcludedPatterns), stream(tsExcludedPatterns))
-              .toArray(WildcardPattern[]::new);
-        }
-      }
-
-      private static boolean isExclusionOverridden(Configuration config) {
-        return (
-          config.get(JavaScriptPlugin.JS_EXCLUSIONS_KEY).isPresent() ||
-            config.get(JavaScriptPlugin.TS_EXCLUSIONS_KEY).isPresent()
-        );
-      }
-
-      @Override
-      public boolean test(Path path) {
-        return WildcardPattern.match(exclusions, path.toString().replaceAll("[\\\\/]", "/"));
-      }
     }
   }
